# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tharaa Al-Maarifa Reading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: batchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a single user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/bulk:
    post:
      operationId: bulkCreateUsers
      tags: [users]
      summary: Bulk create users from pasted text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateUsersBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCreateUsersResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: replaceUser
      tags: [users]
      summary: Full update of user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserFullBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/approve:
    post:
      operationId: approveUser
      tags: [users]
      summary: Approve user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/book:
    patch:
      operationId: updateUserBook
      tags: [users]
      summary: Update user current book and last page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBookBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /batches:
    get:
      operationId: listBatches
      tags: [users]
      summary: List all batches
      responses:
        "200":
          description: Batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [users]
      summary: Create a new batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}:
    put:
      operationId: updateBatch
      tags: [users]
      summary: Update batch name
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBatchBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBatch
      tags: [users]
      summary: Delete batch and all its users and logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        batchId:
          type: ["integer", "null"]
        status:
          type: string
        currentBookId:
          type: ["integer", "null"]
        lastPage:
          type: integer
        phaseNumber:
          type: ["integer", "null"]
        levelType:
          type: ["string", "null"]
        completedBooks:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - status
        - lastPage
        - completedBooks
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        batchId:
          type: ["integer", "null"]
        phaseNumber:
          type: ["integer", "null"]
        levelType:
          type: ["string", "null"]
      required:
        - name
        - phone
        - password
    UpdateUserFullBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        batchId:
          type: ["integer", "null"]
        status:
          type: string
        phaseNumber:
          type: ["integer", "null"]
        levelType:
          type: ["string", "null"]
      required:
        - name
        - phone
        - status
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        batchId:
          type: ["integer", "null"]
        status:
          type: string
        phaseNumber:
          type: ["integer", "null"]
        levelType:
          type: ["string", "null"]
        currentBookId:
          type: ["integer", "null"]
        lastPage:
          type: integer
    UpdateUserBookBody:
      type: object
      properties:
        currentBookId:
          type: ["integer", "null"]
        lastPage:
          type: integer
        completedBookId:
          type: ["integer", "null"]
      required:
        - lastPage
    BulkCreateUsersBody:
      type: object
      properties:
        batchId:
          type: integer
        rawText:
          type: string
        phaseNumber:
          type: integer
        levelType:
          type: string
      required:
        - batchId
        - rawText
        - phaseNumber
        - levelType
    BulkCreateUsersResponse:
      type: object
      properties:
        created:
          type: integer
        failed:
          type: integer
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required:
        - created
        - failed
        - users
    Batch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateBatchBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateBatchBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...