# lib/api-zod/src/generated/types/updateBatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tharaa Al-Maarifa Reading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBatchBody {
  name: string;
}