# lib/api-zod/src/generated/types/updateUserFullBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tharaa Al-Maarifa Reading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateUserFullBody {
  name: string;
  phone: string;
  /** @nullable */
  batchId?: number | null;
  status: string;
  /** @nullable */
  phaseNumber?: number | null;
  /** @nullable */
  levelType?: string | null;
}